Add messages service

+++ src/main/java/server/chat/controller/ChatController.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.*;
 import server.chat.dto.ChatDTO;
 import server.chat.model.Keyword;
 import server.chat.service.ChatService;
+import server.chat.service.MessagesService;
 
 import java.util.List;
 
@@ -15,10 +16,9 @@ public class ChatController {
     @Autowired
     private ChatService chatService;
 
-
     // create chat if not exist
     @GetMapping("/{firstUserId}/{secondUserId}")
-    public int getChatId(@PathVariable String firstUserId, @PathVariable String secondUserId) {
+    public long getChatId(@PathVariable String firstUserId, @PathVariable String secondUserId) {
         return chatService.getChatId(firstUserId, secondUserId);
     }
 
@@ -37,4 +37,5 @@ public class ChatController {
     public void deleteChat(@PathVariable String firstUserId, @PathVariable String secondUserId) {
         chatService.deleteChat(firstUserId, secondUserId);
     }
+
 }
